When I use if - else statemants like this: int var = 20; if (var>10){ System.out.println("First step"); } if (var<21){ System.out.println("Second step"); } else if(var==20){ System.out.println("Third step"); } else{ System.out.println("Fourth step"); } System.out.println("Fifth step");
I will go throw first 2 steps, but then, when I see the right condition (var==20), but there is the "else - if" statement - so I pass this step and go to the Fifth step. But, if there is no else - if, but only if, then I go throw the Third step, and then - to the Fifth step.